+++ lib/screens/map_screen.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+
+class MapScreen extends StatefulWidget {
+  const MapScreen({super.key});
+
+  @override
+  State<StatefulWidget> createState() => MapScreenState();
+}
+
+class MapScreenState extends State<MapScreen> {
+  late final WebViewController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = WebViewController()
+      ..loadRequest(
+        Uri.parse('https://zkmf2024.ch/map'),
+      )
+      ..setJavaScriptMode(JavaScriptMode.unrestricted);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("Karte Festareal"),
+      ),
+      body: Expanded(
+        child: WebViewWidget(
+          controller: _controller,
+        ),
+      ),
+    );
+  }
+}